Move crypto-random-string
