+++ lib/utils/crypto-random-string.js
